Updating the README for the branch for comfort noise.
One bullet point in the todo involved mjordan and file. My memory
from the pizza and drinks session at Astricon is not very clear, but it was when File 
outlined a possible architecture. I need to document that and get him to explain once more,
poor guy. I do hope that he remembers.