Initial elimination of the user@context mailbox format leaking out of app_voicemail.
* Made core not depend upon user@context mailbox format.

* Made channel drivers not depend upon user@context mailbox format.  The
channel drivers just use the given mailbox string as is.

* The chan_dahdi.conf mwi_mailboxes format now associates the internal
mailbox name with a mailbox number to send out on the span.

* The chan_dahdi.conf mwimonitornotify shell script is now passed the
internal mailbox name instead of just the user part of user@context
format.

* Made channel drivers not append @default to a mailbox that does not have
a context.  The exception to this is the users.conf hasvoicemail option.
The option is legacy and could only work with app_voicemail anyway.  The
mailbox option overrides the hasvoicemail option anyway.

* Made the external MWI code not deal with context at all.