stasis: Add eid support to stasis messaging and cache.
* Convert all stasis messages to have an eid to identify its source.

* Make stasis cache store items with the same id but from different source
eids in the same cache entry.

Note: The device state and mwi state events are published to stasis with
the eid source and consumers need to be updated to get an aggregated state
for a specific device and mwi.  The device_state_aggregation_test unit
test currently fails as a result.